Update kdsoap to rev 1 via SR 788069
https://build.opensuse.org/request/show/788069
by user luca_b + dimstar_suse
New package: kdsoap, a C++ Qt library to build client and server SOAP based
components

Compared to the previous requests, this one removes unused and not installed
unit test files with non free licenses from the tarball (see
https://github.com/KDAB/KDSoap/issues/207). As a side effect, GPG verification
of the tarball is temporarily disabled  until upstream fixes the problem.

README.txt has been added to %license to show that the EULA present in the
source is only applicable to the commercial version and not the FOSS
release of this library.