Update zonefs-tools to rev 1 via SR 791786
https://build.opensuse.org/request/show/791786
by user morbidrsa + dimstar_suse
Zonefs is a very simple file system exposing each zone of a zoned block device as a file. Zonefs does not have complex on-disk metadata and relies solely on a super block stored at LBA 0 of the device.

This package contains the user-space tools for zonefs.