Update libplacebo to rev 9 via SR 796364
https://build.opensuse.org/request/show/796364
by user mia + dimstar_suse
- Add libepoxy to BuildRequires
- Enable test suite
- Update to version 2.43.0
The major highlight of this release is the introduction of a new
OpenGL backend, documented in <libplacebo/opengl.h>, which is also
the reason for the major version bump.
Apart from the addition of OpenGL support, this release brings with
it a large number of fixes for backwards compatibility with older
versions of GLSL and GLES, allowing libplacebo to be used to
generate legal shaders for virtually any GLSL version in existence.
Supporting these changes, the test framework has been expanded
considerably to allow for automated testing of older OpenGL context
versions and older GLSL versions.
Additions:
  * add pl_gpu.pci, containing the PCI address of the underlying
    GPU
  * add pl_image.av1_grain, allowing pl