Update cscreen to rev 1 via SR 813172
https://build.opensuse.org/request/show/813172
by user trenn + dimstar_suse
First submitrequest of a screen (screenrc) wrapper

We use this internally for years already.
See NON_Public:BuildOPS or NON_Public:infrastructure.
I polished this up quite a bit (use _cscreen instead of root),
etc. and it would be nice to have this in openSUSE:Factory
for broader feedback and usage.