Update python-sgp4 to rev 1 via SR 820694
https://build.opensuse.org/request/show/820694
by user bnavigator + dimstar_suse
This Python package computes the position and velocity of an earth-orbiting
satellite, given the satellite’s TLE orbital elements from a source like
Celestrak. It implements the most recent version of SGP4, and is regularly run
against the SGP4 test suite to make sure that its satellite position predictions
agree to within 0.1 mm with the predictions of the standard distribution of the
algorithm. This error is far less than the 1–3 km/day by which satellites
themselves deviate from the ideal orbits described in TLE files.

If your platform supports it, this package compiles the verbatim source code
from the official C++ version of SGP4. You can call the routine directly, or
through an array API that loops over arrays of satellites and arrays of times
with machine code instead of Python. Oth