Track version in commit message
this is useful info
and OBS already parses it out for us.
We parse the .rev file in a hacky way
assuming, the last entry is the one we are processing right now.

This avoids an API request to query the rev
or to properly parse the XML
but can lead to some incorrect results in rare cases