Update qutebrowser to version 1.14.1 / rev 47 via SR 853375
https://build.opensuse.org/request/show/853375
by user mia + dimstar_suse
- Update to version 1.14.1
Added
With v1.14.0, qutebrowser configures the main window to be
transparent, so that it's possible to configure a translucent tab-
or statusbar. However, that change introduced various issues, such
as performance degradation on some systems or breaking dmenu window
embedding with its -w option. To avoid those issues for people who
are not using transparency, the default behavior is reverted to
versions before v1.14.0 in this release. A new window.transparent
setting can be set to true to restore the behavior of v1.14.0.
Fixed
* Setting the content.headers.referer setting to same-domain (the
  default) was supposed to truncate referers to only the host with
  QtWebEngine. Unfortunately, this functionality broke in Qt 5.14.
  It works properly again.
* with this re