Update udunits2 to version 2.2.26 / rev 1 via SR 851875
https://build.opensuse.org/request/show/851875
by user manfred999 + dimstar_suse
dsteuer requested to include this package into factory. It is needed by some R packages,
and is used in several scientific software packages (which are potentially-to-be-included-in-factory) as cdo or nco. The compat parts are not really needed, but they just provide some symlinks and do no harm.

Request has been enhanced with splitting off libudunits2-0 as requested
and some minor cleanup.