Update myquerytutor to version 2.1.1+git20201224.0172df6 / rev 1 via SR 860028
https://build.opensuse.org/request/show/860028
by user tuxta + dimstar_suse
New Package. Software that teaches SQL through a series of exercises. Provides instant feedback on the correctness of a query and provides progress graph and topic lessons. I am the author of MyQueryTutor and use it for my high school students. It is also used by many schools in my state.