Update xorgproto to version 2021.4 / rev 6 via SR 889520
https://build.opensuse.org/request/show/889520
by user sndirsch + dimstar_suse
- xorgproto 2021.4
The big new feature in this release is Olivier's addition that makes up
XFixes 6: the new ClientDisconnectMode. An X server that is started on
demand (Xwayland) should ideally also terminate when the last client
disconnects. However, some X11 clients that provide system services will
linger around forever, preventing that shutdown.
With the new XFixes request, a client can designate itself as
to-be-terminated and the X server can ignore those clients when counting the
number of remaining clients. If no other clients are left, the server can
shut down.
Note that this requires changes to the X server and each client to work.

yesterday's tarball.