Update vendor-reset to version 0.1.0+225a49a / rev 1 via SR 924229
https://build.opensuse.org/request/show/924229
by user mrostecki + dimstar_suse
Hi,

This is a new package I would like to maintain. It ships an out-of-tree kernel module which allows to use VFIO for passing through devices which have so called "reset bug" (which means - a device becoming non-functional after the VM, to which it got passed through, shuts down or reboots).

That module is required i.e. for GPUs like Radeon RX 5xxx to be passed through to VMs.

See more on the repo: https://github.com/gnif/vendor-reset