Update monitoring-plugins to version 2.3.1 / rev 23 via SR 933139
https://build.opensuse.org/request/show/933139
by user lrupp + dimstar_suse
JFYI: this is planned as update also for SLE12 and SLE15.
Please have a closer look, not that I missed something.
As the plugins are in productive use on monitor.opensuse.org, I
do not expect big issues. But the dhcpd plugin was not extensively tested, yet.

- recommend syslog for monitoring-plugins-log, as people probably
  want to analize logs generated by (r)syslog or journald
Renamed patches:
- renamed monitoring-plugins-1.4.6-no_chown.patch to
  monitoring-plugins-1.4.6-Makefile_-_no_chown.patch to make it
  easier to detect the patched file
- renamed monitoring-plugins-2.1.1-check_logfile.patch to
  monitoring-plugins-2.1.1-check_log_-_quoting.patch to make it
  easier to detect the patched file and reason for the patch
New patches:
- added monitoring-plugins-2.3.1-check_snmp_segf