Update tpm2-tss-engine to version 1.1.0 / rev 8 via SR 940279
https://build.opensuse.org/request/show/940279
by user mgerstner + dimstar_suse
resulted in non-position-independent executable) (bsc#1183895).

Let's not add dummy changelog entries that makes it hard to understand what is wanted.

All that was intended here was to have the bug ref next to the change responsible to fix it.
so let's do that: add the bug ref to the change responsible for it. (forwarded request 939078 from dimstar)