Update simutrans to version 123.0.1 / rev 5 via SR 952743
https://build.opensuse.org/request/show/952743
by user dstoecker + dimstar_suse
- update to version 123.0.1
- This is a bugfix to 123.0, where the schedule editing did not really work.
- Also pak48, pak128, pak128.german and pak64.german got updates.
- This release is build from svn revision number 10421. This number is configured to allow
network games to be able to connect to an online server running the same stable version.
- Complete history can be found at (unofficial):
  https://github.com/aburch/simutrans/blob/master/simutrans/history.txt
- From now on, the build is configured to use the zstd compression library for saving. Previously
this was done with bzip2 compression, but zstd can replace it. Zstd is faster (advantage)
but also leads to larger files (disadvantage). Usually network games using zstd, since speed is an issue.
So for maximum network gaming c