Delete gnome-gmail-notifier via SR 953064
https://build.opensuse.org/request/show/953064
by user iznogood + dimstar_suse
I have a strong suspicion that this package no longer works as exepect, at least not without jumping quite a few hurdles to get gmail to ack the "plain old imap/pop" support that this package expects.
Futher it depends on old packages we really really want gone from distro, and this is a blocker package for that.