Update hashlink to version 1.11 / rev 1 via SR 974127
https://build.opensuse.org/request/show/974127
by user jaimeMF + dimstar_suse

I want to add HashLink ( https://hashlink.haxe.org/ ) to openSUSE:Factory

HashLink is a virtual machine for Haxe.
Haxe supports two modes of compilation for HashLink:
* Compilation to HashLink bytecode. This mode has a very fast compilation time,
so it is suitable for daily development.
* Compilation to HashLink/C code, compiled with a native compiler to a regular
executable. This mode results in the best performance, so it is suitable for
final releases.