Delete libyami-utils via SR 1001125
https://build.opensuse.org/request/show/1001125
by user iznogood + dimstar_suse
Fails to build, and fails to build with ffmpeg-5 (when tested).
Nothing depends on this package, and upstream have archived the git repo -- see
https://github.com/intel/libyami-utils/commit/3d730dc4da9c29aa0d15539433245b3ab271f971

DISCONTINUATION OF PROJECT.

This project will no longer be maintained by Intel.

Intel has ceased development and contributions including, but not limited to, maintenance, bug fixes, new releases, or updates, to this project.

Intel no longer accepts patches to this project.

If you have an ongoing need to use this project, are interested in independently developing it, or would like to maintain patches for the open source software community, please create your own fork of this project.