Update patterns-devel-vulkan to version 20221115 / rev 1 via SR 1036251
https://build.opensuse.org/request/show/1036251
by user sndirsch + dimstar_suse
Just talked to Frederic Crozat and he thinks it's a good idea to have a "Vulkan Development" a pattern.

Recently I wrote a short blog article [1] about this topic, but of course it's
easier to just install a Vulkan development pattern.

I will open a pull-request on github [2] for the pattern icon, if my SR gets accepted.

[1] https://sndirsch.github.io/vulkan/2022/10/20/vulkan-packages-on-opensuse.html
[2] https://github.com/sndirsch/yast-theme