Update xtrans to version 1.5.0 / rev 17 via SR 1091972
https://build.opensuse.org/request/show/1091972
by user sndirsch + dimstar_suse
- Update to 1.5.0
xtrans is a library of code that is shared among various X packages to handle
network protocol transport in a modular fashion, allowing a single place to
add new transport types - but it is *not* a shared library, more like a \
"header-only" library.  It is used by the X server, the XIM support in libX11,
libICE, the X font server, and related components.  Because this is not a
shared library, the changes in this release will only take effect in consumers
that are rebuilt on a system with this release of xtrans installed.
This release makes progress towards resolving CVE-2020-25697, reported in
https://www.openwall.com/lists/oss-security/2020/11/09/3 .  Clients will no
longer attempt to connect to sockets in the abstract namespace, though
servers will still bind to them to