Update vkbasalt-cli to version 3.1.1 / rev 1 via SR 1093650
https://build.opensuse.org/request/show/1093650
by user VortexAcherontic + dimstar_suse
Hello everyone,

I'd like to submit this package to Factory as it is missing.

About the software:

vkbasalt-cli is a python based command line tool and library to be used with the Vulkan post processing layer vkbasalt.
It's purpose is to make it easier for end-users and applications to generated configuration files for vkbasalt.
vkbasalt-cli is a stable project, hence it is not in any alpha, beta or RC state.

The only software (I know of) making use of vkbasalt-cli under the hood as of now is bottles which on its own is a managing frontend for Wine and related software.
Which then, if vkbasalt-cli can be merged with Factory, can be updated to it's latest version.

I apologize for any errors regarding this request as this is the first time for me submitting to Factory.

Kind regards,
Imo H