Update tpm2-0-tss to version 4.0.1 / rev 33 via SR 1100357
https://build.opensuse.org/request/show/1100357
by user mgerstner + anag+factory
Require openssl-3 over openssl-1 to assist migration of applications to newer openssl-3.

Alternately we might need to have two copies of the package for openssl-3 and openssl-1, but
we have to transition to just openssl-3 at some point.

This is required for Kanidm as it requires openssl-3 and tpm2-0-tss. (forwarded request 1100221 from firstyear)