Update python-pykdumpfile to version 0.5.3.1 / rev 1 via SR 1124285
https://build.opensuse.org/request/show/1124285
by user ptesarik + anag+factory
The Python bindings bundled with libkdumpfile are now deprecated in favour of this package. To prevent regression, submit to Factory before disabling Python build in libkdumpfile.
Version obsolete has been solved with a new upstream release, as long as this SR is accepted together with a libkdumpfile SR.