Delete bin86 via SR 1126240
https://build.opensuse.org/request/show/1126240
by user matz2 + anag+factory
Since the change to dosemu got accepted nothing in Factory
depends on bin86 anymore.  It isn't maintained upstream anymore
and it seems not useful to have it in the large scheme of things.
(It also has a Y2038 problem which was the immediate reason to look
into this).