Delete phonon4qt5-backend-gstreamer via SR 1139342
https://build.opensuse.org/request/show/1139342
by user pallaswept + anag+factory
KDE developers recommend that distros do not include this package as it is buggy, unmaintained and archived.

This version is known to cause the bug https://bugs.kde.org/show_bug.cgi?id=445196 because it does not contain the patch which fixes it.
The version at https://build.opensuse.org/package/show/KDE:Unstable:Frameworks/phonon4qt5-backend-gstreamer contains the patch and fixes the bug. (of course, this is not a solution, as that repository is unstable and cannot be added to the repository list)
The official word from KDE is that distros should not use this, but it was installed by default for me. To quote the KDE developers and maintainers of this package:

"Phonon gstreamer shouldn't be used at all - it hasn't been the recommended backend of phonon for probably half a decade
...Appar