Update blame to version 1.4+t20240206 / rev 1 via SR 1146125
https://build.opensuse.org/request/show/1146125
by user manfred999 + anag+factory
As adding devel projects to the user's installation is discouraged, I'd like to forward
this package to Factory, although the number of users will be quite limited, I guess...

Blame is the RCS equivalent of CVS's "annotate" command. It is quite a common tool
for RCS users.