Update rcshist to version t20230130 / rev 1 via SR 1146127
https://build.opensuse.org/request/show/1146127
by user manfred999 + anag+factory
As adding devel projects to the user's installation is discouraged, I'd like to forward
this package to Factory, although the number of users will be quite limited, I guess...

The rcshist utility displays the complete revision history of an RCS file.
It is quite a common tool for RCS users.