Update collada-dom to version 2.5.1+git20200104.c1e20b7 / rev 1 via SR 1163819
https://build.opensuse.org/request/show/1163819
by user bobbie424242 + anag+factory
I'd like to submit this new package to Factory.

The reason is that it will be needed when the openmw package (that is in Factory) will update to its next major version, probably sometimes in 2024.

Current openmw git (which I build in my home repo) already depends on the COLLADA plugin from the OpenSceneGraph package. This plugin is not currently built because it depends on this new collada-dom package I'm submitting.
If this package is accepted in Factory, I will submit an update to the OpenSceneGraph package so it builds its COLLADA plugin. Then, whenever the next major update of openmw is available it will just be a matter to add a dependency to this plugin to the openmw spec file (all that work is already done in my home repo).