Update python-typer-slim to version 0.12.0 / rev 1 via SR 1164422
https://build.opensuse.org/request/show/1164422
by user theMarix + anag+factory
This is a new package that was split out of python-typer following upstream's split of the source distribution.

This SR will be followed with SRs for the equally split out python-typer-cli, and finally an updated python-typer that will depend on this and python-typer-cli.

I tested the upgrade path and this will not break anybody relying on the package.