Update simutrans to version 124.1 / rev 7 via SR 1178880
https://build.opensuse.org/request/show/1178880
by user m_vanderwulp + anag+factory
- update to version 124.1
- This version mostly fixes bugs reported for previous version 124.0,
with some features added to the top. There are also some changes related to packaging:
1. The 124 release made Freetype a required dependency, but the build system
was not updated to reflect that. As a result, you could compile Simutrans
without Freetype, only for the game to fail at launch because
it could not use any font. We have fixed this for the 124.1 release,
and now Freetype is also required for building (except for server builds).
2. The ".desktop" file has been renamed and moved to "src/linux/simutrans.desktop".
- See for more information about the 124.1 and the 124.0 releases:
https://forum.simutrans.com/index.php/board,3.0.html
- Patch simutrans-makefile.patch added since GNU ma