Delete libevtlog via SR 1216813
https://build.opensuse.org/request/show/1216813
by user czanik + anag+factory
Libevtlog was once a library created for RFC5424 logging. It was designed to be general purpose, but in the end only syslog-ng used it. After many years of separate development, libevtlog was merged into syslog-ng a couple of years ago. It is maintained now as part of syslog-ng, and still can be installed independent of syslog-ng.