Delete Bottles via SR 1234634
https://build.opensuse.org/request/show/1234634
by user VortexAcherontic + anag+factory
Hello there, this is a humble request to remove Bottles from Factory in order to comply with the developers wish not to ship Bottles in distribution repositories.
https://usebottles.com/posts/2022-06-07-an-open-letter/
Shipping Bottles in Factory might create a false view to users and possibly make them assume it is an official package while it isn't.

I do understand that some users do not want to use the bottles Flatpak which is why the package will still be maintained in games:tools. Since games:tools is not setup by default the user will have to actively add the repository and be more aware of it being an unofficial and "experimental" package.

The initial request to add Bottles to Factory was lead by another maintainer who as of the start of this year has stepped down from maintaining