Update wl-mirror to version 0.17.0 / rev 1 via SR 1239787
https://build.opensuse.org/request/show/1239787
by user holzbauer + anag+factory
Hello,
I want to introduce a new package to factory:

wl-mirror is a Wayland output mirror client.
It mirrors (in the sense of copying, not swapping left↔right) an output onto a new resizable window.
https://github.com/Ferdi265/wl-mirror

It is useful for wayland based window managers like hyprland and sway, where no screen mirroring utility is provided by default.
Therefore it would be important to have the package available in the main repos.