Update ncspot to version 1.2.2 / rev 1 via SR 1243266
https://build.opensuse.org/request/show/1243266
by user holzbauer + anag+factory
Hello,
I want to re-introduce a package to factory:

ncspot is a ncurses spotify client written in rust using librespot
https://github.com/hrkfdn/ncspot

It is useful for terminal users, because allows to listen to spotify using only the terminal.
Therefore it would be important to have the package available in the main repos.