Update kmod to version 34.1 / rev 85 via SR 1253001
https://build.opensuse.org/request/show/1253001
by user michals + dimstar_suse
The patch 0001-build-resolve-build-failure-due-to-missing-include.patch is intentionally dropped.

It was added with version 34 update, not mentioned in changelog, and not used.

This addition of patch not mentioned in changelog should have been rejected by the very bot that is rejecting this update but it did not happen.

- tests: drop ppc64 workaround, print failed test results if any