Update extest to version 1+git20241105.1a419a1 / rev 1 via SR 1287216
https://build.opensuse.org/request/show/1287216
by user Tobi_Peter + anag_factory
Initial version of Extest.
Extest is a drop in replacement for the X11 XTEST extension. It creates a virtual device with the uinput kernel module. It's been primarily developed for allowing the desktop functionality on the Steam Controller to work while Steam is open on Wayland.