Update numa-preplace to version 0.1 / rev 1 via SR 1290224
https://build.opensuse.org/request/show/1290224
by user dfaggioli + anag_factory
Once upon a time, there was numad. And there still is, e.g., in Factory, except the upstream project is basically abandoned since something like 2017.

https://build.opensuse.org/package/show/hardware/numad

The package does build, but I've tried installing it in Tumbleweed and using it, and it SEGFAULTs right away in a fraction of a sec! :-O

There are also bugs caused by just having numad enabled, especially if Automatic NUMA balancing (see later) is also on... and that is on by default in all our kernels (and, AFAICT, in all kernels of all distros).

Anyway, these days, there is no need of having something like numad running as a daemon any longer (as, the Linux kernel takes care of that by means of a faeature called Automatic NUMA balancing).

On the other hand, the "pre-placement" of a