Update gamescope-session-steam to version 1+git20250418.7104815 / rev 1 via SR 1297830
https://build.opensuse.org/request/show/1297830
by user Tobi_Peter + dimstar_suse
This package provides the Steam Big Picture Mode session based on gamescope.
This opens Steam in the "SteamOS" mode.

There are linting errors. I removed the polkit rules as they are not needed for the scripts in /usr/bin, they are only placeholders here for Steam to not error out. The bigger issue is the steamos-polkit-helper directory in /usr/bin. This is not allowed in general. But, as far as I'm aware, these paths are hardcoded in Steam, so I can't just move them to /usr/lib. How should I handle them?
I already had an audit about the polkit rules here:
https://bugzilla.opensuse.org/show_bug.cgi?id=1246766