Update extest to version 1+git20241105.1a419a1 / rev 2 via SR 1302018
https://build.opensuse.org/request/show/1302018
by user Tobi_Peter + anag_factory
Steam is a 32bit application, so extest can only work when it's 32bit build is used, even on x86_64 machines. Therefore, this change is required for Steam to actually make use of the library.

I added the baselibs.conf to the sources :)

I added the reason for this change in the changelog