Delete fbi via SR 1300437
https://build.opensuse.org/request/show/1300437
by user tdz + anag_factory
Fbi requires user-space interfaces of the kernel's fbdev subsystem (/dev/fb*). As the fbdev subsystem is obsolete and deprecated, the interfaces have been disabled for improved security as per bsc#1212947. So fbi has not been working on Tumbleweed for a while.

Hence, requesting to remove the package.

https://bugzilla.suse.com/show_bug.cgi?id=1212947