Update xpad-noone to version 0+git20250526.b8e98f3 / rev 1 via SR 1302924
https://build.opensuse.org/request/show/1302924
by user Tobi_Peter + anag_factory
This is the original xpad kernel driver with support for Xbox One controllers removed.
If you are running the xone driver you will have to replace the xpad kernel module with
this one to retain the functionality of Xbox and Xbox 360 controllers.

Added missing copyright notice (sorry)