Update xone-dongle-firmware to version 0.1 / rev 1 via SR 1318010
https://build.opensuse.org/request/show/1318010
by user Tobi_Peter + anag_factory
This package adds a script to download the firmware for Microsoft's Xbox One Controllers. It is required for a specialized driver called "xone". xone is required to support the audio interface of xbox type controllers and its wireless dongle.
Submitting this package to allow installation of the xone package which I will add to regular Factory when this is accepted.

This package is only a download script that presents the TOS of Microsoft during installation. It follows how this package works, the use cases are very similar/identical: https://build.opensuse.org/package/show/openSUSE:Factory/fetchmsttfonts