Update jupiter-dock-updater-bin to version 20250220.02 / rev 1 via SR 1318030
https://build.opensuse.org/request/show/1318030
by user Tobi_Peter + anag_factory
This package provides the dock updater needed to update the official Steam Deck dock.
Another package, jupiter-hw-support, expects the files to land in /usr/lib.

(Request for that package will follow, it is not required for this to work :) )

This package is only a download script that downloads the files.
It follows how this package works, the use cases are very similar/identical: https://build.opensuse.org/package/show/openSUSE:Factory/fetchmsttfonts