Delete steamdeck-firmware via SR 1318013
https://build.opensuse.org/request/show/1318013
by user Tobi_Peter + anag_factory
Hey there,

please check if the license(s) of the distributed files allow it to be in openSUSE:Factory instead of openSUSE:Factory:NonFree. The files redistributed are licensed under these LICENSES, which are not open source licenses, see: https://gitlab.com/evlaV/linux-firmware-neptune/-/blob/jupiter-20250731.1/WHENCE?ref_type=tags

The licenses that apply to the distributed files are:
https://gitlab.com/evlaV/linux-firmware-neptune/-/blob/jupiter-20250731.1/LICENSE.qcom?ref_type=tags
https://gitlab.com/evlaV/linux-firmware-neptune/-/blob/jupiter-20250731.1/LICENSE.QualcommAtheros_ath10k?ref_type=tags

Should this move to the NonFree project?