Update xone to version 0.4.11 / rev 1 via SR 1318347
https://build.opensuse.org/request/show/1318347
by user Tobi_Peter + anag_factory
This package provides the Linux kernel driver for Xbox One and Xbox Series X|S accessories.
it serves as a modern replacement for xpad, aiming to be
compatible with Microsoft's Game Input Protocol (GIP).

Now that xone-dongle-firmware is also in Factory, the driver can easily be installed and used on openSUSE :)