Update hashicorp-serf to version 0.10.2 / rev 1 via SR 1322681
https://build.opensuse.org/request/show/1322681
by user dfaggioli + anag_factory
Serf is a service orchestration and management tool (more info on https://github.com/hashicorp/serf).It offers a decentralized, lightweight, highly available and fault tolerant solution for service discovery and orchestration that runs on Linux, Mac OS-X and Windows.

An efficient and lightweight gossip protocol is used to communicate among the nodes, and node failures are detected and notified to the rest of the cluster. An event system is built on top of Serf, so that events such as deploys, configuration changes, et.c, can be propagated around.
Serf is also completely masterless, meaning there's no single point of failure.

This new package provides the agent that should be started on every node (and a service for running it).It's a cool tool, but besides that, we need it as a dependency